properly renaming all components
