+++ src/components/MaterialUi/Button.component.jsx
@@ -1,7 +1,7 @@
 import Button from '@material-ui/core/Button';
 import React from 'react';
 
-const CustomBtn = (props) => {
+const CustomBtn = props => {
   const { text, onClick, size, variant, color, className } = props;
   return (
     <Button
