+++ src/components/post-listing/post-listing.style.js
@@ -9,20 +9,20 @@ export const StyledContainer = styled.div`
   margin: 0 auto;
   grid-gap: 20px;
 
-  @media ${(props) => props.theme.screen.tablet} {
+  @media ${props => props.theme.screen.tablet} {
     grid-template-columns: repeat(2, 1fr);
   }
 
-  @media ${(props) => props.theme.screen.laptopL} {
+  @media ${props => props.theme.screen.laptopL} {
     grid-template-columns: repeat(3, 1fr);
   }
 `;
 
 export const StyledCard = styled(Card)`
   && {
-    background-color: ${(props) => props.theme.colors.cardItemBg};
-    color: ${(props) => props.theme.colors.bodyTextColor};
-    box-shadow: ${(props) => props.theme.colors.cardItemShadow};
+    background-color: ${props => props.theme.colors.cardItemBg};
+    color: ${props => props.theme.colors.bodyTextColor};
+    box-shadow: ${props => props.theme.colors.cardItemShadow};
   }
 
   .cardActionArea {
@@ -48,7 +48,7 @@ export const StyledCard = styled(Card)`
     }
   }
   .styledLink {
-    color: ${(props) => props.theme.colors.primaryThemeColor};
+    color: ${props => props.theme.colors.primaryThemeColor};
     text-decoration: none !important;
   }
 `;
