+++ src/pages/blog.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Helmet from 'react-helmet';
-import Layout from '../components/Layout/layout.component';
+import Layout from '../components/layout/layout.component';
 import config from '../../data/SiteConfig';
 
 const Blog = () => (
